Moved training to display.py
Set all parameters of the agent and game to the top.

to enable the live game view, set show_games to 1, and set show_every and time_between_displayed_moves accordingly

Set the name of the pickled q_table and the created plot pdf in NAME

set parameters alpha etc.

current reward function uses a dictionary to look reward

Board is set to Englisch, but can be changed to try out Diamond or Triangular Shapes